Add: Introduce new pytest fixtures for RID Manager gateway and use case in test suite
